requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
Pillow==7.1.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
